LayoutTests/http/tests/security/frameNavigation/navigation-ALLOWED-descendant.html

<html>
<head>
<script src="resources/lib.js"></script>
<script>
navUtils.set_frame_count(5);

navUtils.onready = function () {
  navUtils.log('Navigating frames.');
  navUtils.navigateByLocation(frames[0].frames[0]);
  navUtils.navigateByOpen('child1_child0');
  navUtils.navigateByForm('child2_child0');
  navUtils.navigateByHyperlink('child3_child0');
  navUtils.navigateByPlugin('child4_child0');
};

navUtils.ondone = function () {
  navUtils.logWindowCount();
  navUtils.log('Done.');
  if (window.layoutTestController)
    layoutTestController.notifyDone();
};
</script>
</head>
<body>
<div id='console'></div>
<iframe name="child0" src="http://localhost:8000/security/frameNavigation/resources/iframe.html#ready.html"></iframe>
<iframe name="child1" src="http://localhost:8000/security/frameNavigation/resources/iframe.html#ready.html"></iframe>
<iframe name="child2" src="http://localhost:8000/security/frameNavigation/resources/iframe.html#ready.html"></iframe>
<iframe name="child3" src="http://localhost:8000/security/frameNavigation/resources/iframe.html#ready.html"></iframe>
<iframe name="child4" src="http://localhost:8000/security/frameNavigation/resources/iframe.html#ready.html"></iframe>
</body>
</html>
