LayoutTests/http/tests/security/frameNavigation/navigation-ALLOWED-descendant-expected.txt

Navigating frames.
Window count: 1
Done.
    


--------
Frame: 'child0'
--------
localhost:8000


--------
Frame: 'child0_child0'
--------
127.0.0.1

Navigation done.

--------
Frame: 'child1'
--------
localhost:8000


--------
Frame: 'child1_child0'
--------
127.0.0.1

Navigation done.

--------
Frame: 'child2'
--------
localhost:8000


--------
Frame: 'child2_child0'
--------
127.0.0.1

Navigation done.

--------
Frame: 'child3'
--------
localhost:8000


--------
Frame: 'child3_child0'
--------
127.0.0.1

Navigation done.

--------
Frame: 'child4'
--------
localhost:8000


--------
Frame: 'child4_child0'
--------
127.0.0.1

Navigation done.
